fixed remap() type to int

+++ kvmd/tools.py
@@ -34,7 +34,7 @@ from typing import TypeVar
 
 # =====
 def remap(value: int, in_min: int, in_max: int, out_min: int, out_max: int) -> int:
-    return (value - in_min) * (out_max - out_min) // (in_max - in_min) + out_min
+    return int((value - in_min) * (out_max - out_min) // (in_max - in_min) + out_min)
 
 
 # =====
